back_end/requirements.txt
django==1.11.15
djangorestframework==3.8.2
pymysql==0.9.2
django-cors-headers==2.4.0